Several improvements in Lingot Core.
* Improved synchronized access to core results (spectrum and frequency).
* Better definition of interface to Core.
* Moved some functions to lingot-signal.
* Removed warnings.
* Removed 'register' keyword.

+++ test/src/lingot-test-main.c
@@ -101,9 +101,9 @@ int main(void) {
     CU_basic_set_mode(CU_BRM_VERBOSE);
     CU_basic_run_tests();
 
-    int num_failures = CU_get_number_of_failures();
+    unsigned int num_failures = CU_get_number_of_failures();
     CU_cleanup_registry();
 
-    return num_failures;
+    return (int) num_failures;
 
 }
